train-test (group effort)

+++ include/item.h
@@ -22,7 +22,7 @@ class item
     connections common_and_uncommon;
     List specs; /* a list of all of the item's specs, parsed from the json file */
     HashTable words;
-    float *array;
+     double  *array;
 
 public:
     item(string fldr, int i);
@@ -39,8 +39,8 @@ public:
     void set_spec_list(List spec);
     List get_spec_list();
     HashTable get_words_ht();
-    void set_table(float *table);
-    float *get_bow_tfidf();
+    void set_table( double  *table);
+     double  *get_bow_tfidf();
 };
 
 #endif
